Machine/compiler fixes for CLM45 and CISM builds
These changes allow the BLAS and LAPACK libraries to be accessed as part
of an ACME build that includes CLM4.5, or CISM, which are the only ACME
components to use these libraries.  In particular, this allows BG
compsets to build (which require both CLM4.5 and CISM).
Changes:
1) env_mach_specific.mustang: added module load mkl (Intel libraries
containing LAPACK and BLAS) to Mustang-specific setup tasks.
2) config_compilers.xml: added compiler flags to include mkl libraries,
if building on Mustang with Intel.  Particular flag combination was
based on the online Intel Math Kernel Librar Link Line Advisor.