Major refactor of ALM PETSc-based thermal model
This refactor removes the hard coded setup of the ALM's
PETSc based thermal model from the MPP (multi-physics-problem)
library. The setup of MPP problem for ALM now occurs in
the biophysics/mpp/driver/alm directory and setup of other
problems for benchmark and verification will occur in
the biophysics/mpp/driver directory.